+++ src/ts/ui/tutorial/tutorialsPanelContent.ts
